moved common fixtures to common utilities for hq tests
