+++ HQSmokeTests/testCases/conftest.py
@@ -1,24 +1,16 @@
 import os
-from py.xml import html
-import pytest
 
 from configparser import ConfigParser
 from pathlib import Path
-from selenium import webdriver
-from webdriver_manager.chrome import ChromeDriverManager
-from webdriver_manager.firefox import GeckoDriverManager
-
-from common_utilities.path_settings import PathSettings
-from common_utilities.hq_login.login_page import LoginPage
+from common_utilities.fixtures import *
 
 """"This file provides fixture functions for driver initialization"""
-
 
 global driver
 
 
 @pytest.fixture(scope="session")
-def environment_settings():
+def environment_settings_hq():
     """Load settings from os.environ
 
             Names of environment variables:
@@ -33,7 +25,7 @@ def environment_settings():
             """
     settings = {}
     for name in ["url", "login_username", "login_password", "mail_username",
-        "mail_password", "bs_user", "bs_key", "staging_auth_key", "prod_auth_key", "invited_webuser_password"]:
+                 "mail_password", "bs_user", "bs_key", "staging_auth_key", "prod_auth_key", "invited_webuser_password"]:
 
         var = f"DIMAGIQA_{name.upper()}"
         if var in os.environ:
@@ -41,20 +33,21 @@ def environment_settings():
     if "url" not in settings:
         env = os.environ.get("DIMAGIQA_ENV") or "staging"
         subdomain = "www" if env == "production" else env
-        ## updates the url with the project domain while testing in CI
+        # updates the url with the project domain while testing in CI
         project = "a/qa-automation-prod" if env == "production" else "a/qa-automation"
         settings["url"] = f"https://{subdomain}.commcarehq.org/{project}"
     return settings
 
+
 @pytest.fixture(scope="session", autouse=True)
-def settings(environment_settings):
+def settings(environment_settings_hq):
     if os.environ.get("CI") == "true":
-        settings = environment_settings
+        settings = environment_settings_hq
         settings["CI"] = "true"
         if any(x not in settings for x in ["url", "login_username", "login_password",
                                            "mail_username", "mail_password", "bs_user", "bs_key", "staging_auth_key",
                                            "prod_auth_key", "invited_webuser_password"]):
-            lines = environment_settings.__doc__.splitlines()
+            lines = environment_settings_hq.__doc__.splitlines()
             vars_ = "\n  ".join(line.strip() for line in lines if "DIMAGIQA_" in line)
             raise RuntimeError(
                 f"Environment variables not set:\n  {vars_}\n\n"
@@ -71,103 +64,9 @@ def settings(environment_settings):
         )
     settings = ConfigParser()
     settings.read(path)
-    ## updates the url with the project domain while testing in local
+    # updates the url with the project domain while testing in local
     if settings["default"]["url"] == "https://www.commcarehq.org/":
         settings["default"]["url"] = f"{settings['default']['url']}a/qa-automation-prod"
     else:
         settings["default"]["url"] = f"{settings['default']['url']}a/qa-automation"
     return settings["default"]
-
-
-@pytest.fixture(scope="module", autouse=True)
-def driver(settings, browser):
-    web_driver = None
-    chrome_options = webdriver.ChromeOptions()
-    firefox_options = webdriver.FirefoxOptions()
-    if settings.get("CI") == "true":
-        if browser == "chrome":
-            chrome_options.add_argument('--no-sandbox')
-            chrome_options.add_argument('disable-extensions')
-            chrome_options.add_argument('--safebrowsing-disable-download-protection')
-            chrome_options.add_argument('--safebrowsing-disable-extension-blacklist')
-            chrome_options.add_argument('window-size=1920,1080')
-            chrome_options.add_argument("--disable-setuid-sandbox")
-            chrome_options.add_argument('--start-maximized')
-            chrome_options.add_argument('--disable-dev-shm-usage')
-            chrome_options.add_argument('--headless')
-            chrome_options.add_argument("--disable-notifications")
-            chrome_options.add_experimental_option("prefs", {
-                "download.default_directory": str(PathSettings.DOWNLOAD_PATH),
-                "download.prompt_for_download": False,
-                "download.directory_upgrade": True,
-                "safebrowsing.enabled": True})
-        elif browser == "firefox":
-            firefox_options.set_headless()
-            firefox_options.add_argument('--no-sandbox')
-            firefox_options.add_argument('disable-extensions')
-            firefox_options.add_argument('--safebrowsing-disable-download-protection')
-            firefox_options.add_argument('--safebrowsing-disable-extension-blacklist')
-            firefox_options.add_argument('window-size=1920,1080')
-            firefox_options.add_argument("--disable-setuid-sandbox")
-            firefox_options.add_argument('--start-maximized')
-            firefox_options.add_argument('--disable-dev-shm-usage')
-            firefox_options.add_argument('--headless')
-            firefox_options.add_argument("--disable-notifications")
-            firefox_options.set_preference("browser.download.dir", str(PathSettings.DOWNLOAD_PATH))
-    if browser == "chrome":
-        web_driver = webdriver.Chrome(executable_path=ChromeDriverManager().install(), options=chrome_options)
-        print("Chrome version:", web_driver.capabilities['browserVersion'])
-    elif browser == "firefox":
-        web_driver = webdriver.Firefox(executable_path=GeckoDriverManager().install(), options=firefox_options)
-    else:
-        print("Provide valid browser")
-    login = LoginPage(web_driver, settings["url"])
-    login.login(settings["login_username"], settings["login_password"])
-    yield web_driver
-    web_driver.quit()
-
-
-def pytest_addoption(parser):
-    """CLI args which can be used to run the tests with specified values."""
-    parser.addoption("--browser", action="store", default='chrome', choices=['chrome', 'firefox'],
-                     help='Your choice of browser to run tests.')
-
-
-@pytest.fixture(scope="module")
-def browser(request):
-    """Pytest fixture for browser"""
-    return request.config.getoption("--browser")
-
-@pytest.mark.optionalhook
-def pytest_html_results_table_header(cells):
-    #<th class="sortable result initial-sort asc inactive" col="result"><div class="sort-icon">vvv</div>Result</th>
-    cells.insert(1, html.th('Tags', class_="sortable", col="tags"))
-    cells.pop()
-
-@pytest.mark.optionalhook
-def pytest_html_results_table_row(report, cells):
-    cells.insert(1, html.td(report.tags))
-    cells.pop()
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_runtest_makereport(item):
-    pytest_html = item.config.pluginmanager.getplugin("html")
-    outcome = yield
-    report = outcome.get_result()
-    tags = ", ".join([m.name for m in item.iter_markers() if m.name != 'run'])
-    extra = getattr(report, 'extra', [])
-    if report.when == "call" or report.when == "teardown":
-        xfail = hasattr(report, 'wasxfail')
-        if (report.skipped and xfail) or (report.failed and not xfail):
-            print("reports skipped or failed")
-            file_name = report.nodeid.replace("::", "_") + ".png"
-            screen_img = _capture_screenshot(item.funcargs["driver"])
-            if file_name:
-                html = '<div><img src="data:image/png;base64,%s" alt="screenshot" style="width:600px;height:300px;" ' \
-                       'onclick="window.open(this.src)" align="right"/></div>' % screen_img
-                extra.append(pytest_html.extras.html(html))
-        report.extra = extra
-        report.tags = tags
-
-def _capture_screenshot(web_driver):
-    return web_driver.get_screenshot_as_base64()
